requirements.txt
fastapi[all]==0.95.2
mlflow==2.3.2
pandas==2.0.1
pdfplumber==0.9.0
python-dotenv==1.0.0
Requests==2.31.0
uvicorn==0.22.0
transformers==4.29.2
torch==2.0.0